refactor(discord): #4229 S8 — extract emission match arm helpers (move-pure)
Move the watcher_direct_fallback_after_session_bound_ack arm of the
emission match out of the frozen tmux_watcher.rs root into a new sibling
module tmux_watcher/terminal_direct_fallback.rs. Byte-identical move
(reverse-transform diff empty modulo documented mechanical adaptations:
&mut/deref reborrows, should_tag_monitor_origin projection, 13-field
locals destructure). continue/break/return mappings=0; threaded
context=28, locals=13.

tmux_watcher.rs 4039 -> 3559 prod LoC (-480). Ratchets lowered in
hotfile_ratchet.toml + audit_maintainability_giant_baseline.toml; new
file 614 LoC stays under the tmux_watcher/** = 700 namespace cap
(no dedicated entry needed).
